Did a quick fix for segmentation faults that I was getting with my PhysicsManager. Turned out there was pointer invalidation everytime containers were being resized, so members within the chipmunk structs were pointing to old, invalid locations. Am just using the cpNew versions of the functions, so memory is scattered everywhere! Need to map the new versions to the corresponding free rather than destroy funcs.

+++ pyorama/physics/physics_manager.pxd
@@ -72,4 +72,4 @@ cdef class PhysicsManager:
     cpdef void shape_delete(self, Handle shape) except *
     cpdef void shape_set_friction(self, Handle shape, float friction) except *
 
-    cpdef void update(self, double delta) except *
+    cpdef void update(self, float delta) except *
